+++ include/id3/tag.h
@@ -30,7 +30,7 @@
 #define _ID3LIB_TAG_H_
 
 #if defined(__BORLANDC__)
-// due to a bug in borland it sometimes still wants mfc compatibility even when you disable it
+// due to a bug in Borland it sometimes still wants MFC compatibility even when you disable it
 #  if defined(_MSC_VER)
 #    undef _MSC_VER
 #  endif
@@ -126,8 +126,8 @@ public:
   ID3_V2Spec GetSpec() const;
   bool       SetSpec(ID3_V2Spec);
 
-  static size_t IsV2Tag(const uchar*);
-  static size_t IsV2Tag(ID3_Reader&);
+  static uint32 IsV2Tag(const uchar*);
+  static uint32 IsV2Tag(ID3_Reader&);
 
   /* Deprecated! */
   void       AddNewFrame(ID3_Frame* f);
